Refactor api endpoint chat handler to its own function remove legacy `chatWithWorkspace` and cleanup `index.js`
