+++ .vscode/settings.json
@@ -43,6 +43,7 @@
     "searxng",
     "Serper",
     "Serply",
+    "streamable",
     "textgenwebui",
     "togetherai",
     "Unembed",
